add username and password authentication support to client API
